requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
transliterate==1.10.2
whitenoise==5.3.0
